Added VARLIB_PATH and VARLOG_PATH defines to simplify packaging. https://github.com/FrontAccountingERP/FA/commit/4a37a28c49bf900dcc370fd3f21186cedcd632c9
Note: This change in the config file will be needed to be taken into account in any direct deployment by inclusion into RestAPI and integration of FA with other applications for such defines.

`VARLOG_PATH` and `VARLIB_PATH` are defined in the `includes/session.inc` and `install/isession.inc` and will have to be similarly defined in the integration scripts.